Condition for show notifications if connection status is offline

+++ src/services/counters.js
@@ -3,7 +3,8 @@ import {
     postNewCounter,
     incCounter,
     decCounter,
-    deleteCounter
+    deleteCounter,
+    notificationAction,
 } from '../actions';
 
 export const getCountersService = () => {
@@ -20,6 +21,9 @@ export const getCountersService = () => {
 
 export const postNewCounterService = (counterName) => {
     return (dispatch) => {
+        if (navigator.onLine === false) {
+            return dispatch(notificationAction("couldntCreateNewCounter"))
+        }
         fetch('/api/v1/counter', {
             method: 'POST',
             body: JSON.stringify(counterName),
@@ -27,16 +31,22 @@ export const postNewCounterService = (counterName) => {
                 'Content-Type': 'application/json'
             }
         })
-            .then((res) => res.json()
-                .then((res) => {
-                    return dispatch(postNewCounter(res))
-                })
-            )
+            .then((res) => {
+                if (res.ok) {
+                    res.json()
+                        .then((res) => {
+                            return dispatch(postNewCounter(res))
+                        })
+                }
+            })
     }
 }
 
 export const postCounterIncService = (props) => {
     return (dispatch) => {
+        if (navigator.onLine === false) {
+            return dispatch(notificationAction("couldntUpdateCounter"))
+        }
         fetch('/api/v1/counter/inc', {
             method: 'POST',
             body: JSON.stringify(props.props),
@@ -44,16 +54,22 @@ export const postCounterIncService = (props) => {
                 'Content-Type': 'application/json'
             }
         })
-            .then((res) => res.json()
-                .then((res) => {
-                    return dispatch(incCounter(res))
-                })
-            )
+            .then((res) => {
+                if (res.ok) {
+                    res.json()
+                        .then((res) => {
+                            return dispatch(incCounter(res))
+                        })
+                }
+            })
     }
 }
 
 export const postCounterDecService = (props) => {
     return (dispatch) => {
+        if (navigator.onLine === false) {
+            return dispatch(notificationAction("couldntUpdateCounter"))
+        }
         fetch('/api/v1/counter/dec', {
             method: 'POST',
             body: JSON.stringify(props.props),
@@ -61,16 +77,22 @@ export const postCounterDecService = (props) => {
                 'Content-Type': 'application/json'
             }
         })
-            .then((res) => res.json()
-                .then((res) => {
-                    return dispatch(decCounter(res))
-                })
-            )
+            .then((res) => {
+                if (res.ok) {
+                    res.json()
+                        .then((res) => {
+                            return dispatch(decCounter(res))
+                        })
+                }
+            })
     }
 }
 
 export const deleteCounterService = (selectedCounterStore) => {
     return (dispatch) => {
+        if (navigator.onLine === false) {
+            return dispatch(notificationAction("couldntDeleteCounter"))
+        }
         selectedCounterStore.forEach((el) => {
             fetch('/api/v1/counter', {
                 method: 'DELETE',
@@ -79,11 +101,14 @@ export const deleteCounterService = (selectedCounterStore) => {
                     'Content-Type': 'application/json'
                 }
             })
-                .then((res) => res.json()
-                    .then((res) => {
-                        return dispatch(deleteCounter(res))
-                    })
-                )
+                .then((res) => {
+                    if (res.ok) {
+                        res.json()
+                            .then((res) => {
+                                return dispatch(deleteCounter(res))
+                            })
+                    }
+                })
         })
     }
 }
